helpers/ingest-cli/pyproject.toml
[project]
name = "vortex-ingest-cli"
version = "0.1.0"
description = "Walk a directory or codebase, parse files into normalized text, and upload to Vortex Enclave for RAG"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Frank Nguyen", email = "frank.quy.nguyen@gmail.com" }]
readme = "README.md"
keywords = [
    "vortex-enclave",
    "rag",
    "ingest",
    "codebase",
    "knowledge-base",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "vortex-enclave>=0.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pathspec>=0.12.0",
]

[project.scripts]
vortex-ingest = "vortex_ingest_cli.__main__:cli"

[project.optional-dependencies]
docs = [
    # Parsers for richer document formats. All optional — the CLI degrades
    # to text-only mode if these aren't installed.
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "pypdf>=5.0.0",
    "pdfplumber>=0.11.0",
    "trafilatura>=1.12.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
]
all = [
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "pypdf>=5.0.0",
    "pdfplumber>=0.11.0",
    "trafilatura>=1.12.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
]

[project.urls]
Homepage = "https://fusionlab.ai"
Repository = "https://github.com/feepfoop/vortex-enclave-mcp-server"
Issues = "https://github.com/feepfoop/vortex-enclave-mcp-server/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vortex_ingest_cli"]
